+++ src/test/scala/edu/ie3/simona/config/ConfigFailFastSpec.scala
@@ -856,7 +856,7 @@ class ConfigFailFastSpec extends UnitSpec with ConfigTestData {
           intercept[InvalidConfigParameterException] {
             ConfigFailFast invokePrivate checkDataSink(
               Sink(
-                Some(Csv("", "", "", isHierarchic = false)),
+                Some(Csv("", "", "", isHierarchic = false, zipFiles = false)),
                 Some(InfluxDb1x("", 0, "")),
                 None,
               )